ApplicationLibrary/pv/mqtt-pv-infer/requirements.txt
torch
torchvision
torchaudio
paho-mqtt==1.6.1
